\name{DDLeslie_sampler}
\alias{DDLeslie_sampler}
\title{Integrated Population Model for Culling and Aerial Count Data}
\usage{
rcpp_hello()
}
\description{
Returns an \R \code{list} containing the character vector
\code{c("foo", "bar")} and the numeric vector \code{c(0, 1)}.
}
\examples{
rcpp_hello()
}
