\name{eyes}
\alias{eyes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Identity Matrices
}
\description{
Get a n by n identity matrix
}
\usage{
eyes(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
dimension
}
}

\value{
A n by n identity matrix.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
eyes(3)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ identity }% use one of  RShowDoc("KEYWORDS")

