% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luhn.R
\name{recap_luhn}
\alias{recap_luhn}
\title{Calculate important sentences using the Luhn method}
\usage{
recap_luhn(text, n_score_words = 4, stopwords = stopwords::stopwords())
}
\arguments{
\item{text}{A string of length 1 to be summarized.}

\item{n_score_words}{Number of words to be considered when scoring. Defaults to 4.}

\item{stopwords}{A string of words to be ignored when scoring sentences.}
}
\value{
A tibble with 2 columns, sentence and score.
}
\description{
Available package/palette combinations are available in the data.frame
\code{\link[paletteer]{palettes_d_names}}. Both \code{package} and \code{palette}
can be supplied as symbols or strings.
}
